requirements.txt
# Async HTTP client
httpx>=0.27,<1.0

# Web frameworks
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0

# Discord bot
discord.py>=2.7,<3.0

# Configuration
pydantic-settings>=2.2,<3.0

# Database (PostgreSQL)
asyncpg>=0.29,<1.0

# Templates
jinja2>=3.1,<4.0
python-multipart>=0.0.9

# Session signing
itsdangerous>=2.1,<3.0

# Linting
ruff>=0.4

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
